fix: Tear down underlying atoms if atomfs mount fails (#565)
The atomfs bad verity test case ends up leaving verity and loop
devices which then break other test cases.  The root cause is that
when the test case attempts to mount the invalid atomfs the underlying
atom was already mounted.  Once the verity fails, stacker exits but
does not remove any previously mounted at atoms in the molecule.  Fix
this by creating a cleanup function which runs if we fail to mount an
atom.

- Add a check to the atomfs.bats which checks for leftover loop mounts

Fixes: #541

Signed-off-by: Ryan Harper <rharper@woxford.com>