fix: don't break on --substitute values which have commas (#580)
The following:

  stacker build --substitute a=b,c

used to work, until commit 3baba644:

  fix(ci): convert ci is failing due to perms (#439)

This squashed PR included 'chore(go.mod): update cli to
github.com/urfave/cli/v2'.  Switching from urfave/cli/v1 to v2 changes
string slice flag behavior to automatically split on commas.

Luckily there is an app.DisableSliceFlagSeparator flag we can set to
tell it not to do that.  Set that flag.  And add a test for this.

Signed-off-by: Serge Hallyn <shallyn@cisco.com>