fix: cache: serialize legacy import field
Fixes #592 by serializing the `wasLegacyImport` field to ensure that the
cache records the correct value when an image has it set.

Includes a test based on Serge's description in #592.

The symptom is an error "cache miss because layer definition was
changed" when the definition could not have changed since build.
This happens during multi file builds in builds with dependencies, and
in build-then-publish flows, where the build will be fine but the
publish will fail with the same error.

Using `--debug` to show the cache mismatch shows that somewhere along
the line, the layer is being serialized with the wrong default value
when writing to the cache.

Needs to change currentCacheVersion to ensure compatibility, see the
TestCacheEntryChanged function comment for more info.

Signed-off-by: Michael McCracken <mikmccra@cisco.com>