fix: uniquify LXC name to avoid socket clashes
Currently, if:
concurrent stacker runs are happening on a system,
and they are building containers with the same name,
and they are being done inside a mount namespace,
and the path name given as the roots dir is the same,
but the actual mounted volume is different,

then both stackers will be able to acquire the file lock at
$rootdir/.lock, and will go ahead and start containers named $name,
which will then race to set up the lxc control socket, which is named
after the container name and the rootfs path, which are both the same
here.

This fix adds a UUID to the container name, which ensures the sockets do
not clash.

Adds a test 'concurrent.bats' to test this, and includes some other test
changes to make it easier to run tests locally with a cached copy of the
base images.

Signed-off-by: Michael McCracken <mikmccra@cisco.com>