feat: move to atomfs v1.1.3 and golang 1.22
atomfs 1.1.3 has a new API for mounting that changes it to store metadata
about the mounted molecules in a known location. It also cleans up the
unmount code path and added a lot of tests to atomfs.

This moves to the new atomfs version and removes an obsolete mount
option.

atomfs also has a min golang version of 1.22, which we adopt for stacker
here too because 1.21 is EOL at this time. This requires a go mod bump
and tweaks to a few github CI yamls.

note the build.yaml stackerfile now exports GOTOOLCHAIN=auto, which is
the default set in $GOROOT/go.env if you install go via the tgz at
go.dev/dl, but apparently the alpine package doesn't do that. We want
auto, in order to have go respect the `toolchain` directive in go.mod.
This does download a new go toolchain during the build, though, so one
could argue that the right thing here would be to just update the alpine
package, but we can do that at a later time.

It also fixes a typo in atomfs.bats - this test case is also now run in
the atomfs repo itself and the typos were fixed there, but we leave it
here for extra coverage.

Signed-off-by: Michael McCracken <mikmccra@cisco.com>