move to atomfs with fixed umount, update golang to 1.22 (#654)
* feat: move to atomfs v1.1.3 and golang 1.22

atomfs 1.1.3 has a new API for mounting that changes it to store metadata
about the mounted molecules in a known location. It also cleans up the
unmount code path and added a lot of tests to atomfs.

This moves to the new atomfs version and removes an obsolete mount
option.

atomfs also has a min golang version of 1.22, which we adopt for stacker
here too because 1.21 is EOL at this time. This requires a go mod bump
and tweaks to a few github CI yamls.

note the build.yaml stackerfile now exports GOTOOLCHAIN=auto, which is
the default set in $GOROOT/go.env if you install go via the tgz at
go.dev/dl, but apparently the alpine package doesn't do that. We want
auto, in order to have go respect the `toolchain` directive in go.mod.
This does download a new go toolchain during the build, though, so one
could argue that the right thing here would be to just update the alpine
package, but we can do that at a later time.

It also fixes a typo in atomfs.bats - this test case is also now run in
the atomfs repo itself and the typos were fixed there, but we leave it
here for extra coverage.

Signed-off-by: Michael McCracken <mikmccra@cisco.com>

* test: add shellcheck to bats tests, fix lint

A previous test bug would have been caught by linting the bats test
code, so let's do that now.

That error was using an undefined variable in a test. ($last_layer_hash
in atomfs.bats) It had been defined in a different test in the same
file, but was undefined in the test in question. Unfortunately because
of the way bats does bash, shellcheck only identifies that as an INFO
level possible error SC2031.

So this adds a new `lintbats` target that first checks for JUST SC2031,
then checks separately for only error level issues.

This commit also fixes a few error level issues in other files:
- asterisk and whiteout : use glob instead of looping over ls. /shrug
- bom: [ -n "${ZOT_HOST}:${ZOT_PORT}" ] will always be true because of
the colon in there, so this commit eliminates the colon.
- config.bats had a complaint about the indentation of the EOF because
it's in yaml in another here doc? OK then.

Also add the shellcheck package to the build deps.

Signed-off-by: Michael McCracken <mikmccra@cisco.com>

* test: remove dep on zothub.io in squashfs test

this is testing import of squash images then building tar on top, we can
test that without also testing zothub.io's nine fives.

Signed-off-by: Michael McCracken <mikmccra@cisco.com>

* chore: build umoci directly for tests

Instead of using `go get` which now does nothing after moving to 1.22,
or `go install` which fails in github CI, let's just clone and build
umoci ourselves.

Signed-off-by: Michael McCracken <mikmccra@cisco.com>

* test: avoid dockerhub dependency

two tests had bare image names, causing problems with dockerhub rate
limiting. move instead to AWS

also clarifies a skip message which was a little confusing because most
of the test is actually completed.

Signed-off-by: Michael McCracken <mikmccra@cisco.com>

* chore: bump x/net to v0.33.0 for CVE-2024-45337

also bumps the golang toolchain to latest current toolchain as of
right... now

Signed-off-by: Michael McCracken <mikmccra@cisco.com>

---------

Signed-off-by: Michael McCracken <mikmccra@cisco.com>