test: auth.json creds works for image import
This is a test to check existing behavior.

Internally, without stacker needing to pass creds in the opts,
containers/image uses GetCredentials() to look in auth.json for creds
for calls to copy.Image().

this adds a test to cover this case using a zot configured to require auth.
also generates certs for the zot in the test.

some other cleanup in tests

Signed-off-by: Michael McCracken <mikmccra@cisco.com>