fix: use host and path for credentials search
To support different credentials for different paths on a host (e.g. an
artifactory server with multiple repositories), we need stacker to send
the full path to GetCredentials. GetCredentials searches for the full
path, then iterates over subpaths by removing one path component at a
time, so the creds from longest matching subpath are returned.

Updates the import-http-auth test to show this behavior.

Signed-off-by: Michael McCracken <mikmccra@cisco.com>