update use of auth.json credentials for `import`, add test of existing image import behavior (#726)
* ci: ensure we have new skopeo

ensure we get the new skopeo into hack/tools/bin and then put that at
the front of PATH

Signed-off-by: Michael McCracken <mikmccra@cisco.com>

* test: auth.json creds works for image import

This is a test to check existing behavior.

Internally, without stacker needing to pass creds in the opts,
containers/image uses GetCredentials() to look in auth.json for creds
for calls to copy.Image().

this adds a test to cover this case using a zot configured to require auth.
also generates certs for the zot in the test.

some other cleanup in tests

Signed-off-by: Michael McCracken <mikmccra@cisco.com>

* fix: use host and path for credentials search

To support different credentials for different paths on a host (e.g. an
artifactory server with multiple repositories), we need stacker to send
the full path to GetCredentials. GetCredentials searches for the full
path, then iterates over subpaths by removing one path component at a
time, so the creds from longest matching subpath are returned.

Updates the import-http-auth test to show this behavior.

Signed-off-by: Michael McCracken <mikmccra@cisco.com>

* docs: update and consolidate credential info

Signed-off-by: Michael McCracken <mikmccra@cisco.com>

---------

Signed-off-by: Michael McCracken <mikmccra@cisco.com>