test: check handling secrets without leaking
importing a shell file with secrets and sourcing it is the current best
way to use secrets without hardcoding them in the stackerfile and
without leaking them into the image or the process table

using --substitute will result in the secrets ending up in the final
image, because the substituted stackeryaml gets added as an annotation.

This test confirms and documents this behavior, and gives us some
confidence that we won't break this approach in the future
unintentionally.

Signed-off-by: Michael McCracken <mikmccra@cisco.com>