fix: address arm64 build with fixed ppa
The puzzle-os/dev PPA had 2ubuntu8 build of lxc which had files for
amd64, but not for arm64; but the Packages.xz pointed arm64 to 2ubunt8
packages which didn't exist for arm64.

The least worst fix was to create a new PPA with an updated lxc-dev
package build for amd64 and arm64.  Move our setup scripts to this
new puzzle-os/dev-next

Signed-off-by: Ryan Harper <rharper@woxford.com>